requirements.txt
aiohttp
botbuilder-core
botbuilder-schema
colorama
mcp
python-dotenv
agent-framework